Move japicmp reporting section to the profile.
Moving the reporting section to the profile allows a clean opt-in to
JAPIcmp.

The <japicmp.skip> property does not totally disable the japicmp report
if set to false. If japicmp configuration is left in the default
<reporting> section it will show in the site on the reports menu and
displays a blank page for the report.

The <japicmp.skip> property has been removed. The property was only used
to skip the default report. Activation of the profile is using a
profile.japicmp file or using the -P flag on the command line.