Improve reproducibility of generated JARs
Bnd/maven-bundle-plugin adds some extra manifest headers like
"Bnd-LastModified" which record information about the build. This makes it
harder to reproduce the JAR bit for bit on a different machine. The documented
way to remove these is the "-noextraheaders: true" instruction.

Even after applying this instruction, JARs still include a "Include-Resource"
header, whose ordering appears to be file system dependent. It is usually removed by
maven-bundle-plugin by default, but the <_removeheaders> tag in commons-parent
overrides the one specified in maven-plugin-bundle for this purpose. Use
"merged properties", see
https://bnd.bndtools.org/chapters/123-tour-workspace.html, to add additional
values to the property instead of overriding it completely.